Create highlights
In this code, I used double asterisks (**) to highlight the new "Cache" option in the menu. This is just one way to visually distinguish the new option from the others.

Note that this code uses the print() function to output each line of the menu. You can run this code in a Python interpreter or save it to a file and run it as a script.