Create collections
This code defines a simple database of users and their posts, a simple user class, and then creates a user and has them follow some other users. It then defines a function to gather collections (such as posts) from users that the current user is following, and uses this function to get the posts from the users that the current user is following.

Note: This is a very basic example and doesn't include any error handling or security measures that you would need in a real-world application.